+++ cef-web/src/main/java/gov/epa/cef/web/service/dto/ControlPathPollutantDto.java
@@ -1,6 +1,7 @@
 package gov.epa.cef.web.service.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class ControlPathPollutantDto implements Serializable {
 	
@@ -9,7 +10,7 @@ public class ControlPathPollutantDto implements Serializable {
     private Long id;
     private Long controlPathId;
     private PollutantDto pollutant;
-    private Double percentReduction;
+    private BigDecimal percentReduction;
     private Long facilitySiteId;
     
     public Long getId() {
@@ -28,11 +29,11 @@ public class ControlPathPollutantDto implements Serializable {
         this.pollutant = pollutant;
     }
 
-    public Double getPercentReduction() {
+    public BigDecimal getPercentReduction() {
         return percentReduction;
     }
 
-    public void setPercentReduction(Double percentReduction) {
+    public void setPercentReduction(BigDecimal percentReduction) {
         this.percentReduction = percentReduction;
     }
